+++ modules/frontroute/src/main/scala/io/frontroute/Directive.scala
@@ -38,8 +38,8 @@ class Directive[L](
           inner(value)(location, previous, state.leaveDisjunction)
         }(location, previous, state.enterDisjunction)
         .flatMap {
-          case RouteResult.Complete(state, location, result) => Val(RouteResult.Complete(state, location, result))
-          case RouteResult.Rejected                          =>
+          case RouteResult.Matched(state, location, result) => Val(RouteResult.Matched(state, location, result))
+          case RouteResult.Rejected                         =>
             other.tapply { value => (location, previous, state) =>
               inner(value)(location, previous, state.leaveDisjunction)
             }(location, previous, state.enterDisjunction)