+++ modules/frontroute/src/main/scala/io/frontroute/ops/DirectiveOfOptionOps.scala
@@ -5,6 +5,6 @@ import io.frontroute.Directive
 
 class DirectiveOfOptionOps(underlying: Directive[Option[Element]]) {
 
-  @inline def mapOption[R](f: Element => R): Directive[Option[R]] = underlying.map(_.map(f))
+  @inline def optionMap[R](f: Element => R): Directive[Option[R]] = underlying.map(_.map(f))
 
 }