+++ website/src/main/scala/io/frontroute/site/Page.scala
@@ -4,6 +4,7 @@ import com.raquo.laminar.api.L._
 
 final class Page private (
   val path: String,
+  val link: String,
   val title: String,
   val render: () => Element
 )
@@ -12,10 +13,12 @@ object Page {
 
   def apply(
     path: String,
+    link: String,
     title: String,
     render: => Element
   ): Page = new Page(
     path,
+    link,
     title,
     () => render
   )