+++ website/src/main/scala/io/frontroute/site/examples/ex_auth/AuthExample.scala
@@ -49,14 +49,17 @@ object AuthExample
             },
             pathPrefix("private") {
               /* <focus> */
-              requireAuthentication { user => /* </focus> */
+              requireAuthentication { user =>
+                /* </focus> */
                 path("profile") {
                   div(
                     div(cls := "text-2xl", "Profile page."),
                     div(s"User: $user")
                   )
                 }
+                /* <focus> */
               }
+              /* </focus> */
             },
             extractUnmatchedPath { unmatched =>
               div(