+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ### 0,.16.x
 
 * memoize is gone (event streams in routes interfere with `noneMatched`)
-
+* `when` --> `whenTrue` (conflict with laminext)
+* `mapOption` -> `optionMap`
+* 
 ### 0.15.2
 
 Airstream `v0.14.2`.